requirements.txt
phidata==2.6.5
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
python-multipart==0.0.6
pydantic==2.10.3
python-dotenv==1.0.0
docling==2.10.0
duckduckgo-search==6.3.7
groq==0.13.0
newspaper4k==0.9.3.1
lxml_html_clean==0.4.1
lancedb==0.17.0
tantivy==0.22.0